+++ src/Writer.cs
@@ -96,7 +96,7 @@ namespace Dec
             string additionalNote = "";
             if (referenced is Array array && array.Length == 0)
             {
-                additionalNote = " (Note: C# empty arrays often refer to a single shared instance, and it's unclear what Dec should do about this. Come talk to us in Discord if you think you have a good solution. Lists do not have this behvaior.)";
+                additionalNote = " (Note: C# empty arrays often refer to a single shared instance, and it's unclear what Dec should do about this. Come talk to us in Discord if you think you have a good solution. Lists do not have this behavior.)";
             }
 
             if (priorWasShared)
@@ -105,7 +105,7 @@ namespace Dec
             }
             else
             {
-                Dbg.Err($"Attempted to create a new shared reference at [{path.Serialize()}] to an previously-seen unshared object at [{priorPath.Serialize()}]. This may result in an invalid serialization. If this is coming from a Recorder setup, it's likely you either need a .Shared() decorator, or you need to ensure that this object is not serialized elsewhere.{additionalNote}");
+                Dbg.Err($"Attempted to create a new shared reference at [{path.Serialize()}] to a previously-seen unshared object at [{priorPath.Serialize()}]. This may result in an invalid serialization. If this is coming from a Recorder setup, it's likely you either need a .Shared() decorator, or you need to ensure that this object is not serialized elsewhere.{additionalNote}");
             }
         }
     }